feat: add widget to provide COVID data on user's current location
Included a new service that runs in the foreground with a notification
that updates a widget when the user travels to different locations.

The Service will log the postcode of the area which can then be provided to
a government server to then provide that user with COVID data about their
current area.